test(map): split oversized generator and view-builder test files (#3588)
Splits the two 1000+ line integration test files in colonizethis_map into
per-concern files, each well under 500 lines, with no test changes:

- tile_map_generator_test.dart (1036 lines) ->
  tile_map_generator_core_test.dart (orchestration, logging, topology, regression),
  tile_map_generator_resource_rules_test.dart,
  tile_map_generator_terrain_distribution_test.dart,
  tile_map_generator_sea_lakes_test.dart,
  tile_map_generator_seeding_test.dart
- init_game_map_view_builder_test_part1_test.dart (1032 lines) ->
  init_game_map_view_builder_region_data_test.dart,
  init_game_map_view_builder_markers_test.dart,
  init_game_map_view_builder_town_markers_test.dart

All 37 generator tests and 11 view-builder tests preserved verbatim.

Refs #3588
